+++ movies_cinema/src/main/kotlin/mk/ukim/finki/repository/MovieRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface MovieRepository : JpaRepository<Movie, Int> {
+interface MovieRepository : JpaRepository<Movie, Long> {
 }
